enh: refactoring for better code management
