+++ external-import/digintlab-dep/src/dep_connector/__init__.py
@@ -1,0 +1,3 @@
+from dep_connector.connector import DepConnector
+
+__all__ = ["DepConnector"]
